// src/generated/user-api/model/userDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * User Service API
 * User service endpoints (dev docs via springdoc).
 * OpenAPI spec version: v1
 */

export interface UserDto {
  id?: number;
  username?: string;
  name?: string;
  surname?: string;
  email?: string;
  roles?: string[];
  verified?: boolean;
}