// src/generated/user-api/auth-controller/auth-controller.zod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * User Service API
 * User service endpoints (dev docs via springdoc).
 * OpenAPI spec version: v1
 */
import * as zod from 'zod';









export const signupBody = zod.object({
  "username": zod.string().min(1),
  "name": zod.string().min(1),
  "surname": zod.string().min(1),
  "email": zod.string().min(1),
  "password": zod.string().min(1)
})








export const signupWithInvitationBody = zod.object({
  "username": zod.string().min(1),
  "name": zod.string().min(1),
  "surname": zod.string().min(1),
  "invitationToken": zod.string().min(1),
  "password": zod.string().min(1)
})




export const sendPlatformInvitationBody = zod.object({
  "email": zod.string().min(1),
  "roleType": zod.enum(['TEACHER', 'STUDENT', 'ADMIN'])
})




export const resendVerificationEmailBody = zod.object({
  "email": zod.string().min(1)
})




export const requestPasswordResetBody = zod.object({
  "email": zod.string().min(1)
})





export const loginBody = zod.object({
  "username": zod.string().min(1),
  "password": zod.string().min(1)
})

export const changePasswordQueryParams = zod.object({
  "token": zod.string()
})




export const changePasswordBody = zod.object({
  "newPassword": zod.string().min(1)
})

export const confirmEmailQueryParams = zod.object({
  "token": zod.string()
})