// src/generated/user-api/user-controller/user-controller.zod.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * User Service API
 * User service endpoints (dev docs via springdoc).
 * OpenAPI spec version: v1
 */
import * as zod from 'zod';


export const getByIdParams = zod.object({
  "id": zod.number()
})

export const _deleteParams = zod.object({
  "id": zod.number()
})

export const updateParams = zod.object({
  "id": zod.number()
})

export const updateBody = zod.object({
  "id": zod.number().optional(),
  "username": zod.string().optional(),
  "name": zod.string().optional(),
  "surname": zod.string().optional(),
  "email": zod.string().optional(),
  "roles": zod.array(zod.string()).optional(),
  "verified": zod.boolean().optional()
})

export const removeRoleParams = zod.object({
  "id": zod.number(),
  "roleType": zod.enum(['TEACHER', 'STUDENT', 'ADMIN'])
})